#POULPE-305 Check whether group moderates any branch while removing

+++ poulpe-view/poulpe-web-controller/src/main/java/org/jtalks/poulpe/web/controller/group/UserGroupVm.java
@@ -14,15 +14,22 @@
  */
 package org.jtalks.poulpe.web.controller.group;
 
-import javax.annotation.Nonnull;
-
+import org.hibernate.Query;
+import org.hibernate.SessionFactory;
+import org.hibernate.classic.Session;
 import org.jtalks.common.model.entity.Group;
+import org.jtalks.poulpe.model.entity.PoulpeBranch;
 import org.jtalks.poulpe.service.GroupService;
 import org.jtalks.poulpe.web.controller.SelectedEntity;
 import org.jtalks.poulpe.web.controller.WindowManager;
+import org.jtalks.poulpe.web.controller.zkutils.BindUtilsWrapper;
 import org.zkoss.bind.annotation.Command;
 import org.zkoss.bind.annotation.NotifyChange;
+import org.zkoss.zul.Combobox;
 import org.zkoss.zul.ListModelList;
+
+import javax.annotation.Nonnull;
+import java.util.List;
 
 /**
  * View-model for 'User Groups' Is used to order to work with page that allows admin to manage groups(add, edit,
@@ -31,20 +38,25 @@ import org.zkoss.zul.ListModelList;
  * @author Leonid Kazancev
  */
 public class UserGroupVm {
+    @SuppressWarnings("JpaQlInspection")
     private static final String SHOW_DELETE_DIALOG = "showDeleteDialog", SHOW_GROUP_DIALOG = "showGroupDialog",
-            SELECTED_GROUP = "selectedGroup";
+            SELECTED_GROUP = "selectedGroup", MODERAING_GROUP_ID = "moderating_group_id",
+            QUERY = "from PoulpeBranch where MODERATORS_GROUP_ID=:", SHOW_MODERATING_BRANCHES = "showModeratingBranches",
+            MODERATING_BRANCHES = "moderatingBranches";
+
 
     //Injected
     private GroupService groupService;
     private final WindowManager windowManager;
+    private final SessionFactory sessionFactory;
 
     private ListModelList<Group> groups;
     private Group selectedGroup;
     private SelectedEntity<Group> selectedEntity;
     private String searchString = "";
+    private BindUtilsWrapper bindWrapper = new BindUtilsWrapper();
 
-    private boolean showDeleteDialog;
-    private boolean showGroupDialog;
+    private boolean showDeleteDialog, showGroupDialog, showModeratingBranches;
 
     /**
      * Construct View-Model for 'User groups' view.
@@ -52,12 +64,14 @@ public class UserGroupVm {
      * @param groupService   the group service instance
      * @param selectedEntity the selected entity instance
      * @param windowManager  the window manager instance
+     * @param sessionFactory the session factory instance
      */
     public UserGroupVm(@Nonnull GroupService groupService, @Nonnull SelectedEntity<Group> selectedEntity,
-                       @Nonnull WindowManager windowManager) {
+                       @Nonnull WindowManager windowManager, @Nonnull SessionFactory sessionFactory) {
         this.groupService = groupService;
         this.selectedEntity = selectedEntity;
         this.windowManager = windowManager;
+        this.sessionFactory = sessionFactory;
 
         this.groups = new ListModelList<Group>(groupService.getAll(), true);
     }
@@ -91,14 +105,19 @@ public class UserGroupVm {
     }
 
     /**
-     * Deletes selected group.
+     * Deletes selected group. Opens branch dialog if group to delete moderating some branches.
      */
     @Command
-    @NotifyChange({SELECTED_GROUP, SHOW_DELETE_DIALOG})
+    @NotifyChange({SELECTED_GROUP, SHOW_DELETE_DIALOG, SHOW_MODERATING_BRANCHES})
     public void deleteGroup() {
-        groupService.deleteGroup(selectedGroup);
-        closeDialog();
-        updateView();
+        if (hasModeratingGroup()) {
+            showModeratingBranches = true;
+            bindWrapper.postNotifyChange(UserGroupVm.this, MODERATING_BRANCHES);
+        } else {
+            groupService.deleteGroup(selectedGroup);
+            updateView();
+            closeDialog();
+        }
     }
 
     /**
@@ -117,14 +136,12 @@ public class UserGroupVm {
     @Command
     @NotifyChange({SELECTED_GROUP, SHOW_GROUP_DIALOG})
     public void showEditDialog() {
-         showGroupDialog = true;
+        showGroupDialog = true;
     }
 
     /**
      * Saves group, closing group edit(add) dialog and updates view.
-     *
      */
-
     @Command
     @NotifyChange({SHOW_GROUP_DIALOG})
     public void saveGroup() {
@@ -137,11 +154,38 @@ public class UserGroupVm {
      * Close all dialogs by set visibility to false.
      */
     @Command
-    @NotifyChange({SHOW_GROUP_DIALOG, SHOW_DELETE_DIALOG})
+    @NotifyChange({SHOW_GROUP_DIALOG, SHOW_DELETE_DIALOG, SHOW_MODERATING_BRANCHES})
     public void closeDialog() {
-
         showDeleteDialog = false;
         showGroupDialog = false;
+        showModeratingBranches = false;
+    }
+
+    /**
+     * @return list of branches moderated by selected group
+     */
+    @SuppressWarnings("unchecked")
+    public List<PoulpeBranch> getModeratingBranches() {
+        if (selectedGroup != null) {
+            getSession().flush();
+            Query query = getSession().createQuery(QUERY + MODERAING_GROUP_ID);
+            query.setParameter(MODERAING_GROUP_ID, selectedGroup.getId());
+            return query.list();
+        } else return null;
+    }
+     
+    /**
+     * @return current Session instance
+     */
+    private Session getSession(){
+        return sessionFactory.getCurrentSession();
+    }
+
+    /**
+     * @return true if group are moderating some branches
+     */
+    private boolean hasModeratingGroup() {
+        return getModeratingBranches().size() != 0;
     }
 
     // -- Getters/Setters --------------------
@@ -149,7 +193,7 @@ public class UserGroupVm {
     /**
      * Gets visibility status of Delete dialog window.
      *
-     * @return true if dialog is visible false if dialog is invisible
+     * @return true if dialog is visible
      */
     public boolean isShowDeleteDialog() {
         return showDeleteDialog;
@@ -160,12 +204,26 @@ public class UserGroupVm {
      * which don't send anything to the server when closing window because of event.stopPropagation, so during next
      * change notification ZK will think that we need to show that dialog again which is wrong.
      *
-     * @return true if dialog is visible false if dialog is invisible
+     * @return true if dialog is visible
      */
     public boolean isShowGroupDialog() {
         boolean show = showGroupDialog;
         showGroupDialog = false;
         return show;
+    }
+
+    /**
+     * @return true if show branches dialog is visible
+     */
+    public boolean isShowModeratingBranches() {
+        return showModeratingBranches;
+    }
+
+    /**
+     * @return groupService instance for creating comboboxes.
+     */
+    public GroupService getGroupService() {
+        return groupService;
     }
 
     /**
@@ -212,5 +270,12 @@ public class UserGroupVm {
      */
     public void setSearchString(String searchString) {
         this.searchString = searchString;
+    }
+
+    /**
+     * @param bindWrapper instance of bindWrapper
+     */
+    public void setBindWrapper(BindUtilsWrapper bindWrapper) {
+        this.bindWrapper = bindWrapper;
     }
 }
